Move files into consolidated Doctrine plugin
