+++ src/LiteCQRS/Plugin/Doctrine/CouchDBTransactionalHandlerFactory.php
