+++ src/LiteCQRS/Plugin/Doctrine/EventStore/CouchDBEventStore.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace LiteCQRS\Plugin\DoctrineCouchDB\EventStore;
+namespace LiteCQRS\Plugin\Doctrine\EventStore;
 
 use LiteCQRS\EventStore\EventStoreInterface;
 use LiteCQRS\EventStore\SerializerInterface;
